smbus: Add support to force open a device
Add two more functions:
- OpenFileWithOptions
- OpenWithOptions

and reimplement the existing ones on top of the introduced functions.

For now the options only allow to pass the 'Force' flag to force open
a device currently under kernel control. This is useful to see if a
kernel driver is running and to inspect registers when writing kernel
drivers for I2C/Smbus devices.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>